* renames UsersComponentBloc to UsersComponentPresenter * renames references
