+++ src/app/users/container/presenter/component.html
@@ -1,4 +1,4 @@
-<ng-container *ngIf="bloc.usernames$ | async as usernames">
+<ng-container *ngIf="presenter.usernames$ | async as usernames">
   <mat-list>
     <mat-list-item class="usernames__element" *ngFor="let username of usernames">{{username}}</mat-list-item>
   </mat-list>
