+++ src/app/users/container/presenter/presenter.spec.ts
@@ -3,14 +3,14 @@ import { createUsersFixture } from '@app/common/users/fixture';
 import { createUsersRestAdapterMock } from '@app/common/users/mock';
 import { UsersService } from '@app/common/users/service';
 import { of } from 'rxjs';
-import { UsersComponentBloc } from './bloc';
+import { UsersComponentPresenter } from './presenter';
 
-describe('UsersComponentBloc', () => {
+describe('UsersComponentPresenter', () => {
   const users = createUsersFixture();
   const userNames = users.map(user => user.name);
   const usersRestAdapterMock = createUsersRestAdapterMock();
   let usersService: UsersService;
-  let bloc: UsersComponentBloc;
+  let presenter: UsersComponentPresenter;
 
   beforeEach(() => {
     usersRestAdapterMock.list.mockReset();
@@ -19,13 +19,13 @@ describe('UsersComponentBloc', () => {
 
   beforeEach(() => {
     usersService = new UsersService(usersRestAdapterMock);
-    bloc = new UsersComponentBloc(usersService);
+    presenter = new UsersComponentPresenter(usersService);
   });
 
   describe('usernames$', () => {
     it('should emit list of user names', () => {
       createRxTestScheduler().run(({ expectObservable }) => {
-        expectObservable(bloc.usernames$).toBe('a', { a: userNames });
+        expectObservable(presenter.usernames$).toBe('a', { a: userNames });
       });
     });
     it('should emit empty list in case there are no users', () => {
@@ -33,7 +33,7 @@ describe('UsersComponentBloc', () => {
         usersRestAdapterMock.list.mockReturnValue(of([]));
         usersService.refreshUsers();
 
-        expectObservable(bloc.usernames$).toBe('a', { a: [] });
+        expectObservable(presenter.usernames$).toBe('a', { a: [] });
       });
     });
   });
@@ -41,7 +41,7 @@ describe('UsersComponentBloc', () => {
     it('should trigger refresh of user list in UsersService', () => {
       usersService = new UsersService(usersRestAdapterMock);
       const refreshUsersSpy = spyOn(usersService, 'refreshUsers');
-      bloc = new UsersComponentBloc(usersService);
+      presenter = new UsersComponentPresenter(usersService);
 
       expect(refreshUsersSpy).toHaveBeenCalled();
     });
