+++ src/app/users/module.ts
@@ -2,9 +2,9 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { MatListModule } from '@angular/material/list';
 import { MatTabsModule } from '@angular/material/tabs';
-import { UsersBlocComponent } from './container/bloc/component';
 import { UsersComponent } from './container/component/component';
 import { UsersMenuComponent } from './container/menu/component';
+import { UsersBlocComponent } from './container/presenter/component';
 import { UsersRoutingModule } from './routing.module';
 
 @NgModule({
