+++ src/app/users/routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { UsersBlocComponent } from './container/bloc/component';
 import { UsersComponent } from './container/component/component';
 import { UsersMenuComponent } from './container/menu/component';
+import { UsersBlocComponent } from './container/presenter/component';
 
 const routes: Routes = [
   {
